started fixing -Wsign-compare warning messages
Import internal commit e7788e63409831d2c4106d462a3205f9625bcda4

+++ groups/bsl/bslstl/bslstl_set.t.cpp
@@ -1434,13 +1434,13 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
 
             Obj mY(X);  const Obj& Y = mY;
 
             ASSERTV(LINE, 0 == verifyContainer(Y, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
 
             Obj mZ;  const Obj& Z = mZ;
 
@@ -1448,7 +1448,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
 
             ASSERTV(LINE, 0 == verifyContainer(Z, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    2 * TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    2 * TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -1458,7 +1458,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
             mX.insert(BEGIN, END);
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         CONT.resetIterators();
@@ -1472,7 +1472,7 @@ void TestDriver<KEY, COMP, ALLOC>::testCase22()
             }
             ASSERTV(LINE, 0 == verifyContainer(X, EXP, LENGTH));
             ASSERTV(LINE, da.numBlocksInUse(),
-                    TYPE_ALLOC * LENGTH == da.numBlocksInUse());
+                    TYPE_ALLOC * LENGTH == (size_t) da.numBlocksInUse());
         }
 
         ASSERTV(LINE, da.numBlocksInUse(), 0 == da.numBlocksInUse());
